Check for redundant image arrays, and skip them.
In issue 76 and pull 72, two of the three files are such that both the figure
co-ordinate list and the image per page are wrong: the former because of
garbage at the end of page data, the latter because of having N * N redundant
images.

At least for the latter, we could do a quick check the image size array
looks like N repeats of a vector size N, to get a correct value.